+++ lib/include/arch_sse_type1.h
@@ -360,17 +360,17 @@ flush_job_snow3g_uia2_sse(MB_MGR_SNOW3G_OOO *state);
 
 /* SM4 */
 void
-sm4_ecb_sse(const void *in, void *out, const int size, const void *exp_keys);
+sm4_ecb_sse(const void *in, void *out, const uint64_t size, const void *exp_keys);
 
 void
-sm4_cbc_enc_sse(const void *in, void *out, const int size, const void *exp_enc_keys,
+sm4_cbc_enc_sse(const void *in, void *out, const uint64_t size, const void *exp_enc_keys,
                 const void *iv);
 void
-sm4_cbc_dec_sse(const void *in, void *out, const int size, const void *exp_dec_keys,
+sm4_cbc_dec_sse(const void *in, void *out, const uint64_t size, const void *exp_dec_keys,
                 const void *iv);
 void
-sm4_cntr_sse(const void *in, void *out, const int size, const void *exp_enc_keys, const void *iv,
-             const int iv_len);
+sm4_cntr_sse(const void *in, void *out, const uint64_t size, const void *exp_enc_keys,
+             const void *iv, const uint64_t iv_len);
 
 void
 sm4_set_key_sse(const void *pKey, void *exp_enc_keys, void *exp_dec_keys);
